Se agregó configuración de recycler view con llamada a retrofit y mappers
