+++ app/src/main/java/com/mariods/harrypotterapp/data/Repository.kt
@@ -1,13 +1,18 @@
 package com.mariods.harrypotterapp.data
 
 import com.mariods.harrypotterapp.data.network.BooksService
-import com.mariods.harrypotterapp.data.network.model.BooksModel
+import com.mariods.harrypotterapp.domain.model.BooksList
+import com.mariods.harrypotterapp.domain.model.toDomain
+
 import javax.inject.Inject
 
 class Repository @Inject constructor(private val booksService: BooksService) {
 
-    suspend fun getAllBooksFromApi(): BooksModel{
-        return booksService.getAllBooks()
+    suspend fun getAllBooksFromApi(): List<BooksList> {
+        val response = booksService.getAllBooks()
+        return response.data.map {
+            it.toDomain()
+        }
     }
 
 }
