+++ app/src/main/java/com/mariods/harrypotterapp/data/network/model/BooksModel.kt
@@ -23,6 +23,7 @@ data class Attributes(
 
 
 
+
 /*
 {
     "data": [
