+++ app/src/main/java/com/mariods/harrypotterapp/domain/usecases/GetAllBooks.kt
@@ -1,12 +1,13 @@
 package com.mariods.harrypotterapp.domain.usecases
 
 import com.mariods.harrypotterapp.data.Repository
-import com.mariods.harrypotterapp.data.network.model.BooksModel
+import com.mariods.harrypotterapp.domain.model.BooksList
+import com.mariods.harrypotterapp.domain.model.toDomain
 import javax.inject.Inject
 
 class GetAllBooks @Inject constructor(private val repository: Repository) {
 
-    suspend operator fun invoke() : BooksModel {
+    suspend operator fun invoke() : List<BooksList> {
         val allBooks = repository.getAllBooksFromApi()
         return allBooks
     }
